Add ParcaScrapeConfig
This commit adds a ParcaScrapeConfig Custom Resource Definition to the
Parca Operator. The CRD can be used to create the "scrape_configs"
section in the Parca configuration file via CRs.

A user must specify a selector for Pods in the CR, which is then used to
generate the targets for the scrape configuration. This way the scrape
configuration can be updated dynamically. All other values in the scrape
configuration can also be specified in the CR (e.g. scrape interval,
scrape timeout, etc.).