Rename Configuration Environment Variables
This commit renames the existing environment variables used to configure
the Parca Operator. The environment variables now follow the following
schema: "PARCA_CONTROLLER-NAME_XYZ". This should improve the
configuration of the operator when we add additional controllers.

This commit also introduces a new environment variable
"PARCA_SCRAPECONFIG_RECONCILIATION_INTERVAL" which can be used to set
the reconciliation interval for the ParcaScrapeConfigs CRs. By default
the will be reconciled every 5 minutes.

The Helm chart version was also updated with the adjusted environment
variables and prepared for the v0.2.0 release of the operator.