Refactor ParcaScrapeConfig CRD and Logic
This commit refactors the CRD schema for the `ParcaScrapeConfig` to be
more similar to the CRDs known from the Prometheus Operator /
VictoriaMetrics.

We also have adjusted the logic of the created Parca configuration.
Instead of getting all the Pods for the specified selector in the
`ParcaScrapeConfig` and adding the Pods as static configs, we are now
using the Kubernetes Service Discovery feature of Parca and just add the
corresponding configuration.